Move Cell package and update relevant affected files
