+++ src/main/java/seedu/address/model/Model.java
@@ -3,7 +3,6 @@ package seedu.address.model;
 import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.cell.CellMap;
 import seedu.address.model.person.Person;
 import seedu.address.model.cell.Cell;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
